move protocol, utils and version into common
