+++ contrib/opencensus-ext-azure/opencensus/ext/azure/common/protocol.py
