+++ contrib/opencensus-ext-azure/opencensus/ext/azure/common/version.py
